All this world is not 32bits...

+++ src/s_conf.c
@@ -19,7 +19,7 @@
  *
  *  (C) 1988 University of Oulu,Computing Center and Jarkko Oikarinen"
  *
- *  $Id: s_conf.c,v 7.239 2001/06/11 19:56:51 leeh Exp $
+ *  $Id: s_conf.c,v 7.240 2001/06/12 21:43:59 androsyn Exp $
  */
 
 #include <sys/types.h>
@@ -107,7 +107,7 @@ static void clear_special_conf(struct ConfItem **);
 typedef struct ip_entry
 {
 #ifndef IPV6
-  unsigned long ip;
+  u_int32_t ip;
 #else
   struct irc_inaddr ip;
 #endif
@@ -769,7 +769,7 @@ static int hash_ip(struct irc_inaddr *addr)
 {
 #ifndef IPV6
   int hash;
-  unsigned long ip;
+  u_int32_t ip;
 
   ip = ntohl(PIN_ADDR(addr));
   hash = ((ip >> 12) + ip) & (IP_HASH_SIZE-1);
